check loadtime>0 to test whether indexing is finished
TBC: Under certain circumstances a server node may return load_pct=100
even though indexing is not yet completed; subsequent queries using the
index fail with error code 203 ("index not readable"). It seems we can
use loadtime as an additional indicator to check whether the index
operation is complete, but need to confirm with the server team.