Add add'l timeout to prevent test failures
The client sends the truncate info command to a single cluster node; the
cluster distributes the command to the remaining nodes asynchronously.
The truncate command may return before the operation has been completed.
Therefore we need to give the cluster some time to delete all the
records before we count the remaining records to determine the test
status.